requirements.txt
streamlit
PyPDF2
python-dotenv
google-generativeai
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
sentence-transformers
faiss-cpu
